r19013: Fix some 64-bit warnings.
Jelmer, Jerry, in cmd_unixinfo.c:159 there is still the warning that "info" is
being used uninitialized which I don't really know how to fix.

Any hints?

Volker
(This used to be commit 4a5d11e194ebdfb6c21586d1cf902f8ed2478e45)

+++ source3/rpcclient/cmd_lsarpc.c
@@ -63,9 +63,9 @@ static void display_query_info_1(DOM_QUERY_1 d)
 {
 	d_printf("percent_full:\t%d\n", d.percent_full);
 	d_printf("log_size:\t%d\n", d.log_size);
-	d_printf("retention_time:\t%lld\n", d.retention_time);
+	d_printf("retention_time:\t%lld\n", (long long)d.retention_time);
 	d_printf("shutdown_in_progress:\t%d\n", d.shutdown_in_progress);
-	d_printf("time_to_shutdown:\t%lld\n", d.time_to_shutdown);
+	d_printf("time_to_shutdown:\t%lld\n", (long long)d.time_to_shutdown);
 	d_printf("next_audit_record:\t%d\n", d.next_audit_record);
 	d_printf("unknown:\t%d\n", d.unknown);
 }
